pyproject.toml
[project]
name = "vidsynth-ai"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ffmpeg>=1.4",
    "loguru>=0.7.3",
    "manim>=0.19.0",
    "matplotlib>=3.10.0",
    "networkx>=3.4.2",
    "openai>=1.65.2",
    "python-dotenv>=1.0.1",
    "pyyaml>=6.0.2",
    "regex>=2024.11.6",
    "tree-sitter-languages>=1.10.2",
    "tree-sitter-python>=0.23.6",
]
